charcaters are appearing and can be selected by race
