--- src/views/Characters/Charatcers.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import { useEffect, useState } from 'react';
-import { fetchCharacters } from '../../services/characters';
-export default function Characters() {
-  const [characters, setCharacters] = useState([]);
-  useEffect(() => {
-    const fetchData = async () => {
-      const data = await fetchCharacters();
-      setCharacters(data);
-    };
-    fetchData();
-  }, []);
-
-  return (
-    <div>
-      <h1>Characters</h1>
-    </div>
-  );
-}
